로직 개선 구현중. 적 사라지는 오류 해결중

+++ JavaGUiGame/src/abstracts/PlayerPositionProvider.java
@@ -1,6 +1,6 @@
 package abstracts;
 
 public interface PlayerPositionProvider {
-    int getPlayerX();
-    int getPlayerY();
+    double getPlayerX();
+    double getPlayerY();
 }
